split script
